feat: refactor logging functionality by moving default_stream_handler to logger module and updating imports
